requirements.txt
matplotlib=3.7.0
networkx=3.1
numpy=1.23.5
pytorch=1.13.1
scipy=1.10.0
torchaudio=0.13.1
torchvision=0.14.1
tqdm=4.64.1

